Use type checking in tests
This adds adds type annotations to tests and fixes all errors.

(This change also runs Mypy in CI to avoid regressions.)

Practically, add ‘-> None’ annotations to unit tests. This tells mypy to
check these tests. This is a good thing since most unit tests are
essentially a ‘public API consumer’.

This revealed a number of issues, which can also be seen by passing
--check-untyped-defs on the version of the code before this commit:

    $ mypy --check-untyped-defs result/tests.py
    [...]
    Found 39 errors in 1 file (checked 1 source file)

The first category is simple and happens many times, namely missing
annotations:

    result/tests.py: note: In function "test_ok_factories":
    result/tests.py:9:1: error: Function is missing a return type annotation
    result/tests.py:9:1: note: Use "-> None" if function does not return a value

The second category are missing annotations that inference cannot solve.
For example, the type of ‘x = Ok()’ is unknown because it's a generic
class. The fix is to annotate using ‘x: Result[None, None] = Ok()’.

The third category are mostly mypy limitations on type inference:

    result/tests.py: note: In function "test_map_or":
    result/tests.py:128:1: error: Function is missing a return type annotation
    result/tests.py:132:12: error: Cannot infer type argument 1 of "map_or" of "Err"
    result/tests.py:132:38: error: Unsupported left operand type for + ("object")
    result/tests.py:132:38: error: Returning Any from function declared to return "str"
    result/tests.py:137:12: error: Cannot infer type argument 1 of "map_or" of "Err"

These are fixed by avoiding unnecessary lambdas. For example, ‘lambda x:
str(x)’ is the same as just ‘str’, so use that. Other example: instead of
a lambda to ‘x+x’ use ‘str.upper’, then adapt the expected values
accordingly.

Finally, suppress flake8 errors in the pattern matching tests.