Migrate to a modern (mostly) declarative setuptools config (#63)
This migrates the packaging infra to a modern mostly declarative
setuptools configuration, almost literally following the best practices
layed out in the setuptools documentation; see https://setuptools.pypa.io/

The version info has moved to a ‘result.__version__’ attribute (in the
‘__init__.py’ file), which is a common pattern in many packages, and
setuptools reads it from there (single source of truth). Run black+isort
on the file while at it.

The new packages can be built the modern way using the PEP517 compliant
‘build’ tool via ‘python -m build’ after a ‘pip install build’. Reflect
this in RELEASING.md. The remaining setup.py file is almost empty, but
it is not removed because it is still needed for --editable development
installs. See setuptools docs and this article for details:
https://blog.ganssle.io/articles/2021/10/setup-py-deprecated.html

Also added a packaging build step to the CI to ensure packaging keeps working.